Script: move STACK extraction into "stack.py"
